+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog for NeoFS Node
 - Do not panic with bad inputs for `GET_RANGE` (#2007)
 - Correctly select the shard for applying tree service operations (#1996)
 - Physical child object removal by GC (#1699)
+- Increase error counter for write-cache flush errors (#1818)
 
 ### Removed
 ### Updated